New York Yankees Weather Project
In this project, I wanted to see how the weather affected the 2014-2015 New York Yankees. I took 3 datasets from Baseball Reference and Kaggle to create this project. I had 2 datasets from Baseball Reference and I combined them and did data cleansing in Excel to make it easier for me. From there I joined the baseball dataset to a New York City weather dataset from Kaggle together in SQL to analyze how the weather affected the New York Yankees. I looked at the Wins versus Losses conditioning on different things dealing with weather like temperature, rain, and if the game was played at day or night. From this project, I found that the Yankees tended to play better when it was warmer, when it rained, and at night.